+++ euphonic_sqw_models/euphonic_wrapper.py
@@ -1,6 +1,7 @@
 import warnings
 from typing import Optional, Dict, Union, Tuple
 
+import psutil
 import numpy as np
 
 from euphonic import (ForceConstants, QpointPhononModes, DebyeWaller,
@@ -51,9 +52,12 @@ class CoherentCrystal(object):
     conversion_mat : (3, 3) float ndarray or None (default: None)
         This matrix is applied to the input (hkl) q-point values before
         the phonon calculation
-    chunk : float (default: 5000)
-        If non-zero will chunk the phonon calculation into blocks of
-        <chunk> q-points
+    chunk : int (default: None)
+        Will chunk the phonon calculation into blocks of
+        <chunk> q-points. If not provided, an optimum chunk
+        size will be calculated based on available memory. If
+        a chunk size of zero is provided, the calculation will
+        not be chunked.
     lim : float (default: infinity)
         A cut-off for the calculated structure factor where values
         above `lim` are set equal to it
@@ -79,7 +83,7 @@ class CoherentCrystal(object):
             bose: bool = True,
             negative_e: bool = False,
             conversion_mat: Optional[np.ndarray] = None,
-            chunk: int = 5000,
+            chunk: int = None,
             lim: float = np.inf,
             scattering_lengths: Union[
                 str, Dict[str, Union[Quantity, float]]] = 'Sears1992',
@@ -112,9 +116,12 @@ class CoherentCrystal(object):
         conversion_mat
             Shape (3, 3) float ndarray. This matrix is applied to the
             input (hkl) q-point values before the phonon calculation
-        chunk
-            If non-zero will chunk the phonon calculation into blocks of
-            <chunk> q-points
+        chunk : int (default: None)
+            Will chunk the phonon calculation into blocks of
+            <chunk> q-points. If not provided, an optimum chunk
+            size will be calculated based on available memory. If
+            a chunk size of zero is provided, the calculation will
+            not be chunked.
         lim
             A cut-off for the calculated structure factor where values
             above `lim` are set equal to it
@@ -140,6 +147,9 @@ class CoherentCrystal(object):
         self.bose = bose
         self.negative_e = negative_e
         self.conversion_mat = conversion_mat
+        if chunk is None:
+            chunk = self.get_optimum_chunk_size(
+                force_constants.crystal.n_atoms)
         self.chunk = chunk
         self.lim = lim
         self.scattering_lengths = scattering_lengths
@@ -261,6 +271,14 @@ class CoherentCrystal(object):
         dw_phonons = self._calculate_phonon_modes(dw_qpts)
         self.debye_waller = dw_phonons.calculate_debye_waller(
             self.temperature)
+
+    @staticmethod
+    def get_optimum_chunk_size(n_atoms: int):
+        mem = psutil.virtual_memory().available
+        evec_bytes_per_qpt = 16*(3*n_atoms)**2
+        # Divide chunk by 10 for a conservative estimate
+        chunk = int(mem/(10*evec_bytes_per_qpt))
+        return chunk
 
     @property
     def force_constants(self) -> ForceConstants: